+++ src/layouts/DialogLayout/partial/TheAuthDialog/partial/LogInForm/validations/logIn.js
@@ -1,0 +1,17 @@
+import { email, required } from '@vuelidate/validators'
+import { getPasswordValidator } from '@validators'
+
+export default function () {
+  return {
+    logInForm: {
+      email: {
+        email,
+        required
+      },
+      password: {
+        password: getPasswordValidator(),
+        required,
+      }
+    }
+  }
+}
