+++ src/views/User/Contacts/validations/filledContacts.js
@@ -1,0 +1,7 @@
+export default function () {
+  return {
+    filledContacts: {
+      atLeastOne: ( val ) => Object.keys( val ).length >= 1
+    }
+  }
+}
