+++ src/views/User/Me/Me.vue
@@ -196,8 +196,7 @@
 
 <script>
 import useVuelidate from '@vuelidate/core'
-import { required, between, minLength, url, numeric } from '@vuelidate/validators'
-import { phone } from '@validators'
+import userFormValidation from './validations/userForm'
 
 import { getObjectPartClone } from '@functions'
 import { getBackendFormatDate } from '@filters'
@@ -287,52 +286,7 @@ export default {
       this.userForm = getObjectPartClone( userData, keys )
     }
   },
-  validations () {
-    return {
-      userForm: {
-        firstName: {
-          required,
-        },
-        lastName: {
-          required,
-        },
-        street: {
-          required,
-        },
-        city: {
-          required
-        },
-        country: {
-          required
-        },
-        // zipcode: {
-        //   required,
-        // },
-        phoneNumber: {
-          phone
-        },
-        instagramUrl: {
-          required,
-          url
-        },
-        height: {
-          required,
-          minLength: minLength( 3 ),
-          between: between( 99, 250 )
-        },
-        bustImplants: {
-          required,
-          isBustType: ( val ) => Object.entries( this.bustType ).some( ( type ) => type[ 1 ].id === val )
-        },
-        contactServices: {
-        },
-        age: {
-          numeric,
-          between: between( 18, 99 )
-        },
-      }
-    }
-  },
+  validations: userFormValidation()
 }
 </script>
